Added first page for recovery keychain
