+++ catalyst_voices/apps/voices/integration_test/pageobject/onboarding/create_flow/step_11_seedphrase_success.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../utils/translations_utils.dart';
-import '../common_page.dart';
-import 'onboarding_base_page.dart';
+import '../../../utils/translations_utils.dart';
+import '../../common_page.dart';
+import '../onboarding_base_page.dart';
 import 'step_10_input_seedphrase.dart';
 
 class SeedphraseSuccessPanel extends OnboardingPageBase {
