+++ catalyst_voices/apps/voices/integration_test/pageobject/onboarding/create_flow/step_16_wallet_list.dart
@@ -1,24 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../utils/translations_utils.dart';
-import '../common_page.dart';
-import 'onboarding_base_page.dart';
-import 'step_14_keychain_final.dart';
+import '../../../utils/translations_utils.dart';
+import '../../common_page.dart';
+import '../onboarding_base_page.dart';
+import 'step_15_link_wallet_info.dart';
 
-class LinkWalletInfoPanel extends OnboardingPageBase {
-  LinkWalletInfoPanel(super.$);
-
-  static const chooseCardanoWalletButton = Key('ChooseCardanoWalletButton');
-
-  Future<void> clickChooseCardanoWallet() async {
-    await $(chooseCardanoWalletButton).tap();
-  }
+class WalletListPanel extends OnboardingPageBase {
+  WalletListPanel(super.$);
 
   @override
   Future<void> goto() async {
-    await KeychainFinalPanel($).goto();
-    await KeychainFinalPanel($).clickLinkWalletAndRoles();
+    await LinkWalletInfoPanel($).goto();
+    await LinkWalletInfoPanel($).clickChooseCardanoWallet();
   }
 
   @override
@@ -26,8 +20,6 @@ class LinkWalletInfoPanel extends OnboardingPageBase {
     await verifyInfoPanel();
     await verifyDetailsPanel();
   }
-
-  Future<void> verifyDetailsPanel() async {}
 
   Future<void> verifyInfoPanel() async {
     expect(
@@ -46,4 +38,6 @@ class LinkWalletInfoPanel extends OnboardingPageBase {
       T.get('Learn More'),
     );
   }
+
+  Future<void> verifyDetailsPanel() async {}
 }
