+++ catalyst_voices/apps/voices/integration_test/pageobject/onboarding/create_flow/step_3_setup_base_profile.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:patrol_finders/patrol_finders.dart';
 
-import '../../utils/translations_utils.dart';
-import 'onboarding_base_page.dart';
+import '../../../utils/translations_utils.dart';
+import '../onboarding_base_page.dart';
 import 'step_2_base_profile_info.dart';
 
 class SetupBaseProfilePanel extends OnboardingPageBase {
