+++ catalyst_voices/apps/voices/integration_test/pageobject/onboarding/create_flow/step_6_catalyst_keychain_info.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../utils/translations_utils.dart';
-import '../common_page.dart';
-import 'onboarding_base_page.dart';
+import '../../../utils/translations_utils.dart';
+import '../../common_page.dart';
+import '../onboarding_base_page.dart';
 import 'step_5_base_profile_final.dart';
 
 class CatalystKeychainInfoPanel extends OnboardingPageBase {
