+++ catalyst_voices/apps/voices/integration_test/pageobject/onboarding/step_1_get_started.dart
@@ -12,6 +12,11 @@ class GetStartedPanel extends OnboardingPageBase {
   final createNewKeychain = const Key('CreateAccountType.createNew');
   final recoverKeychain = const Key('CreateAccountType.recover');
 
+  @override
+  Future<void> goto() async {
+    await $(AppBarPage.getStartedBtn).tap(settleTimeout: Time.short.duration);
+  }
+
   Future<void> clickCreateNewKeychain() async {
     await $(createNewKeychain).tap();
   }
